feat!(viewmodel): move the editorial reductions out of zingolib into zingo-viewmodel
The value-transfer derivation and everything downstream of it leave
zingolib: the ValueTransfer type family with its Display and JSON
renderings, the self-send value zeroing, messages_containing with its
empty-memo filtering and text search, the three do_total_* finsight
rollups, and the chain-generic fixture that exercises them. They now
live in zingo-viewmodel behind two extension traits,
LightWalletViewModelExt and LightClientViewModelExt, which preserve the
exact method names and signatures the wallet and light client carried,
so a consumer migrates by adding the dependency and one `use` — the
diff zingo-cli takes in this same commit. The golden harness proves the
move: the pre-extraction goldens reproduce byte-for-byte against the
relocated implementation.

zingolib keeps the canonical layer — transaction_summaries, the
note/coin summaries, and classification — and promotes three lossless
accessors (pools_present, shielded_notes_by_pool, received_memos) to
pub so the editorial crate derives from public canonical data like any
consumer. The editorial unit tests, the zero_value_receipts mock-chain
twin, and the create_various_value_transfers fixture move with the
logic; libtonode-tests re-points its imports. This implements the
boundary rule of the pending ADR 0013: a reduction stays in zingolib
only when it needs privilege AND admits no reasonable disagreement.
