+++ zingo-cli/Cargo.toml
@@ -12,6 +12,7 @@ nym = ["zingolib/nym"]
 pepper-sync = { workspace = true }
 zingo_common_components = { workspace = true }
 zingo-netutils = { workspace = true }
+zingo-viewmodel = { workspace = true }
 zingolib = { workspace = true }
 
 zcash_address = { workspace = true }
