+++ zingo-viewmodel/Cargo.toml
@@ -8,21 +8,27 @@ repository = "https://github.com/zingolabs/zingolib"
 homepage = "https://github.com/zingolabs/zingolib"
 license = "MIT"
 
+[features]
+# Chain-generic fixtures for the editorial layer, for integration-test
+# crates that drive them against a LocalNet.
+testutils = ["zingolib/testutils", "dep:tracing"]
+
 [dependencies]
-# Intentionally empty at the scaffold stage: the editorial layer moves here
-# in the extraction increments, at which point zingolib becomes a real
-# dependency and the `nym = ["zingolib/nym"]` feature lands with the
-# mixnet-status view.
+chrono = { workspace = true }
+json = { workspace = true }
+tracing = { workspace = true, optional = true }
+zcash_protocol = { workspace = true }
+zingo-status = { workspace = true }
+zingolib = { workspace = true }
 
 [dev-dependencies]
-json = { workspace = true }
+incrementalmerkletree = { workspace = true }
+orchard = { workspace = true }
 pepper-sync = { workspace = true, features = ["test-features"] }
 tokio = { workspace = true, features = ["full"] }
 zcash_client_backend = { workspace = true }
 zcash_primitives = { workspace = true }
-zcash_protocol = { workspace = true }
 zingo_common_components = { workspace = true }
-zingo-status = { workspace = true }
 zingo_test_vectors = { workspace = true }
 zingolib = { workspace = true, features = ["testutils"] }
 zip32 = { workspace = true }
